Split the cpu_features
